feat: np indexer removes unused indices

+++ jina/executors/indexers/vector.py
@@ -26,27 +26,30 @@ class BaseNumpyIndexer(BaseVectorIndexer):
         Setting :attr:`compress_level`>0 gives a smaller file size on the disk in the index time. However, in the query
         time it loads all data into memory at once. Not ideal for large scale application.
 
-        Setting :attr:`compress_level`=0 enables :func:`np.memmap`, which loads data in an on-demanding way and
+        Setting :attr:`compress_level`=0 enables :func:`np.memmap`, which loads data in an on-demand way and
         gives smaller memory footprint in the query time. However, it often gives larger file size on the disk.
+
+    :param compress_level: The compresslevel argument is an integer from 0 to 9 controlling the
+                    level of compression; 1 is fastest and produces the least compression,
+                    and 9 is slowest and produces the most compression. 0 is no compression
+                    at all. The default is 9.
+    :param ref_indexer: Bootstrap the current indexer from a ``ref_indexer``. This enables user to switch
+                        the query algorithm at the query time.
+    :param delete_on_dump: whether to delete the rows marked as delete (see ``valid_indices``)
     """
 
     def __init__(self,
                  compress_level: int = 1,
                  ref_indexer: Optional['BaseNumpyIndexer'] = None,
+                 delete_on_dump: bool = False,
                  *args, **kwargs):
-        """
-        :param compress_level: The compresslevel argument is an integer from 0 to 9 controlling the
-                        level of compression; 1 is fastest and produces the least compression,
-                        and 9 is slowest and produces the most compression. 0 is no compression
-                        at all. The default is 9.
-        :param ref_indexer: Bootstrap the current indexer from a ``ref_indexer``. This enables user to switch
-                            the query algorithm at the query time.
-        """
         super().__init__(*args, **kwargs)
         self.num_dim = None
         self.dtype = None
+        self.delete_on_dump = delete_on_dump
         self.compress_level = compress_level
         self.key_bytes = b''
+        self.valid_indices = np.array([], dtype=bool)
         self.ref_indexer_workspace_name = None
 
         if ref_indexer:
@@ -59,6 +62,7 @@ class BaseNumpyIndexer(BaseVectorIndexer):
             self._size = ref_indexer._size
             # point to the ref_indexer.index_filename
             # so that later in `post_init()` it will load from the referred index_filename
+            self.valid_indices = ref_indexer.valid_indices
             self.index_filename = ref_indexer.index_filename
             self.logger.warning(f'\n'
                                 f'num_dim extracted from `ref_indexer` to {ref_indexer.num_dim} \n'
@@ -67,23 +71,66 @@ class BaseNumpyIndexer(BaseVectorIndexer):
                                 f'compress_level overriden from `ref_indexer` to {ref_indexer.compress_level} \n'
                                 f'index_filename overriden from `ref_indexer` to {ref_indexer.index_filename}')
             self.ref_indexer_workspace_name = ref_indexer.workspace_name
-
-    def post_init(self):
-        super().post_init()
-        self.valid_indices = np.array([], dtype=bool)
+            self.delete_on_dump = getattr(ref_indexer, 'delete_on_dump', delete_on_dump)
 
     def _delete_invalid_indices(self):
-        """Here rebuild the index with only valid entries"""
-        pass
+        valid = self.valid_indices[self.valid_indices == True]  # noqa
+        if len(valid) != len(self.valid_indices):
+            self._clean_memmap()
+            self._post_clean_memmap(valid)
+
+    def _post_clean_memmap(self, valid):
+        # here we need to make sure the fields
+        # that depend on the valid_indices are cleaned up too
+        valid_key_bytes = np.frombuffer(self.key_bytes, dtype=(np.str_, self.key_length))[
+            self.valid_indices].tobytes()
+        self.key_bytes = valid_key_bytes
+        self._size = len(valid)
+        self.valid_indices = valid
+        del self.__dict__['CACHED__int2ext_id']
+        del self.__dict__['CACHED__ext2int_id']
+
+    def _clean_memmap(self):
+        # clean up the underlying matrix of entries marked for deletion
+        # first we need to make sure we flush the writing handler
+        if self.write_handler and not self.write_handler.closed:
+            with self.write_handler as f:
+                f.flush()
+            self.handler_mutex = False
+        # force the raw_ndarray (underlying matrix) to re-read from disk
+        # (needed when there were writing ops to be flushed)
+        del self.__dict__['CACHED__raw_ndarray']
+        filtered = self._raw_ndarray[self.valid_indices]
+        # we need an intermediary file
+        tmp_path = self.index_abspath + 'tmp'
+
+        # write the bytes in the respective files
+        if self.compress_level > 0:
+            with gzip.open(tmp_path, 'wb', compresslevel=self.compress_level) as new_gzip_fh:
+                new_gzip_fh.write(filtered.tobytes())
+        else:
+            with open(tmp_path, 'wb') as filtered_data_fh:
+                filtered_data_fh.write(filtered.tobytes())
+
+        os.remove(self.index_abspath)
+        os.rename(tmp_path, self.index_abspath)
+        # force it to re-read again from the new file
+        del self.__dict__['CACHED__raw_ndarray']
 
     def __getstate__(self):
-        self._delete_invalid_indices()
+        # called on pickle save
+        if self.delete_on_dump:
+            self._delete_invalid_indices()
         d = super().__getstate__()
         return d
 
     @property
     def workspace_name(self):
-        """Get the workspace name."""
+        """Get the workspace name.
+
+
+        .. # noqa: DAR201
+        """
         return self.name if self.ref_indexer_workspace_name is None else self.ref_indexer_workspace_name
 
     @property
@@ -92,6 +139,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
 
         Use index_abspath
 
+
+        .. # noqa: DAR201
         """
         return self.get_file_from_workspace(self.index_filename)
 
@@ -137,6 +186,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
 
         :param keys: a list of ``id``, i.e. ``doc.id`` in protobuf
         :param vectors: embeddings
+        :param args: not used
+        :param kwargs: not used
         """
         np_keys = np.array(keys, (np.str_, self.key_length))
         self._add(np_keys, vectors)
@@ -154,6 +205,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
 
         :param keys: a list of ``id``, i.e. ``doc.id`` in protobuf
         :param vectors: embeddings
+        :param args: not used
+        :param kwargs: not used
         """
         # noinspection PyTypeChecker
         if self.size:
@@ -176,6 +229,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
         """Delete the embeddings from the index via document ids.
 
         :param keys: a list of ``id``, i.e. ``doc.id`` in protobuf
+        :param args: not used
+        :param kwargs: not used
         """
         if self.size:
             keys = self._filter_nonexistent_keys(keys, self._ext2int_id.keys())
@@ -199,17 +254,20 @@ class BaseNumpyIndexer(BaseVectorIndexer):
             return self.build_advanced_index(vecs)
 
     def build_advanced_index(self, vecs: 'np.ndarray'):
-        """Build advanced index structure based on in-memory numpy ndarray, e.g. graph, tree, etc.
+        """Not implemented here.
 
-        :param vecs: the raw numpy ndarray
-        :return:
+
+        .. # noqa: DAR201
+
+
+        .. # noqa: DAR101
         """
         raise NotImplementedError
 
-    def _load_gzip(self, abspath: str) -> Optional['np.ndarray']:
+    def _load_gzip(self, abspath: str, mode='rb') -> Optional['np.ndarray']:
         try:
             self.logger.info(f'loading index from {abspath}...')
-            with gzip.open(abspath, 'rb') as fp:
+            with gzip.open(abspath, mode) as fp:
                 return np.frombuffer(fp.read(), dtype=self.dtype).reshape([-1, self.num_dim])
         except EOFError:
             self.logger.error(
@@ -224,8 +282,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
             return self._load_gzip(self.index_abspath)
         elif self.size is not None and os.stat(self.index_abspath).st_size:
             self.logger.success(f'memmap is enabled for {self.index_abspath}')
-            deleted_keys = len(self.valid_indices[self.valid_indices == False])
             # `==` is required. `is False` does not work in np
+            deleted_keys = len(self.valid_indices[self.valid_indices == False])  # noqa
             return np.memmap(self.index_abspath, dtype=self.dtype, mode='r',
                              shape=(self.size + deleted_keys, self.num_dim))
 
@@ -234,6 +292,8 @@ class BaseNumpyIndexer(BaseVectorIndexer):
         Search the index by the external key (passed during `.add(`).
 
         :param keys: a list of ``id``, i.e. ``doc.id`` in protobuf
+        :param args: not used
+        :param kwargs: not used
         :return: ndarray of vectors
         """
         keys = self._filter_nonexistent_keys(keys, self._ext2int_id.keys())
@@ -245,7 +305,11 @@ class BaseNumpyIndexer(BaseVectorIndexer):
 
     @cached_property
     def _int2ext_id(self) -> Optional['np.ndarray']:
-        """Convert internal ids (0,1,2,3,4,...) to external ids (random index) """
+        """Convert internal ids (0,1,2,3,4,...) to external ids (random strings)
+
+
+        .. # noqa: DAR201
+        """
         if self.key_bytes:
             r = np.frombuffer(self.key_bytes, dtype=(np.str_, self.key_length))
             # `==` is required. `is False` does not work in np
@@ -253,6 +317,9 @@ class BaseNumpyIndexer(BaseVectorIndexer):
             if r.shape[0] == (self.size + deleted_keys) == self._raw_ndarray.shape[0]:
                 return r
             else:
+                print(f'the size of the keys and vectors are inconsistent '
+                      f'({r.shape[0]}, {self._size}, {self._raw_ndarray.shape[0]}), '
+                      f'did you write to this index twice? or did you forget to save indexer?')
                 self.logger.error(
                     f'the size of the keys and vectors are inconsistent '
                     f'({r.shape[0]}, {self._size}, {self._raw_ndarray.shape[0]}), '
@@ -260,7 +327,11 @@ class BaseNumpyIndexer(BaseVectorIndexer):
 
     @cached_property
     def _ext2int_id(self) -> Optional[Dict]:
-        """Convert external ids (random index) to internal ids (0,1,2,3,4,...) """
+        """Convert external ids (random strings) to internal ids (0,1,2,3,4,...)
+
+
+        .. # noqa: DAR201
+        """
         if self._int2ext_id is not None:
             return {k: idx for idx, k in enumerate(self._int2ext_id)}
 
@@ -301,26 +372,25 @@ def _cosine(A_norm_ext, B_norm_ext):
 
 
 class NumpyIndexer(BaseNumpyIndexer):
-    """An exhaustive vector indexers implemented with numpy and scipy. """
+    """An exhaustive vector indexers implemented with numpy and scipy.
 
+    .. note::
+        Metrics other than `cosine` and `euclidean` requires ``scipy`` installed.
+
+    :param metric: The distance metric to use. `braycurtis`, `canberra`, `chebyshev`, `cityblock`, `correlation`,
+                    `cosine`, `dice`, `euclidean`, `hamming`, `jaccard`, `jensenshannon`, `kulsinski`,
+                    `mahalanobis`,
+                    `matching`, `minkowski`, `rogerstanimoto`, `russellrao`, `seuclidean`, `sokalmichener`,
+                    `sokalsneath`, `sqeuclidean`, `wminkowski`, `yule`.
+    :param backend: `numpy` or `scipy`, `numpy` only supports `euclidean` and `cosine` distance
+    :param compress_level: compression level to use
+    """
     batch_size = 512
 
     def __init__(self, metric: str = 'cosine',
                  backend: str = 'numpy',
                  compress_level: int = 0,
                  *args, **kwargs):
-        """
-        :param metric: The distance metric to use. `braycurtis`, `canberra`, `chebyshev`, `cityblock`, `correlation`,
-                        `cosine`, `dice`, `euclidean`, `hamming`, `jaccard`, `jensenshannon`, `kulsinski`,
-                        `mahalanobis`,
-                        `matching`, `minkowski`, `rogerstanimoto`, `russellrao`, `seuclidean`, `sokalmichener`,
-                        `sokalsneath`, `sqeuclidean`, `wminkowski`, `yule`.
-        :param backend: `numpy` or `scipy`, `numpy` only supports `euclidean` and `cosine` distance
-
-        .. note::
-            Metrics other than `cosine` and `euclidean` requires ``scipy`` installed.
-
-        """
         super().__init__(*args, compress_level=compress_level, **kwargs)
         self.metric = metric
         self.backend = backend
@@ -332,6 +402,9 @@ class NumpyIndexer(BaseNumpyIndexer):
         Idea is to use partial sort to retrieve top-k smallest distances unsorted and then sort these
         in ascending order. Equivalent to full sort but faster for n >> k. If k >= n revert to full sort.
 
+        :param dist: the distances
+        :param top_k: nr to limit
+        :return: tuple of indices, computed distances
         """
         if top_k >= dist.shape[1]:
             idx = dist.argsort(axis=1)[:, :top_k]
@@ -356,6 +429,11 @@ class NumpyIndexer(BaseNumpyIndexer):
 
             Distance (the smaller the better) is returned, not the score.
 
+        :param vectors: the vectors with which to search
+        :param *args: not used
+        :param **kwargs: not used
+        :param top_k: nr of results to return
+        :return: tuple of indices within matrix and distances
         """
         if self.size == 0:
             return np.array([]), np.array([])
